+++ SistemaPlanificacion.Entity/Presupuesto.cs
@@ -19,7 +19,7 @@ public partial class Presupuesto
 
     public DateTime? FechaPresupuesto { get; set; }
 
-    public float? MontoPresupuesto { get; set; }
+    public decimal? MontoPresupuesto { get; set; }
 
     public bool? Nulo { get; set; }
